refactor components into separate files, remove bullet points
