+++ src/App.css
@@ -2,3 +2,7 @@
   text-align: center;
 }
 
+ul {
+  list-style: none;
+}
+
